refactor: add custom hook useTimerparamsFromSearchParams
